requirements.txt
flask
gunicorn
opencv-python
mediapipe
numpy
pandas
PyQt5
rerun-sdk
